% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_param.R
\name{check_param}
\alias{check_param}
\title{check_param}
\usage{
check_param(param_list, rv)
}
\arguments{
\item{param_list}{Param list}

\item{rv}{Reactive value}
}
\value{
Check param
}
\description{
check_param
}
