% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_param.R
\name{import_param_list}
\alias{import_param_list}
\title{import_param_list}
\usage{
import_param_list(module, rv, session)
}
\arguments{
\item{module}{Module name}

\item{rv}{Reactivevalue list}

\item{session}{Shiny session}
}
\value{
A param list
}
\description{
import_param_list
}
