% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load_dendro.R
\name{load_dendro}
\alias{load_dendro}
\title{load_dendrometric_datas}
\usage{
load_dendro(rv, session)
}
\arguments{
\item{rv}{Reactive values}

\item{session}{Session}
}
\description{
load_dendrometric_datas
}
