% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load_sup_datas.R
\name{load_sup_datas}
\alias{load_sup_datas}
\alias{load_parcellaire}
\title{load_supplementary_datas}
\usage{
load_parcellaire(rv, session)
}
\arguments{
\item{rv}{Reactive values}

\item{session}{Session}
}
\description{
load_supplementary_datas
}
