% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cumul.R
\name{mod_cumul_ui}
\alias{mod_cumul_ui}
\alias{mod_cumul_server}
\title{mod_cumul_ui and mod_cumul_server}
\usage{
mod_cumul_ui(id)

mod_cumul_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
