% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dendro.R
\name{mod_dendro_ui}
\alias{mod_dendro_ui}
\alias{mod_dendro_server}
\title{mod_dendro_ui and mod_dendro_server}
\usage{
mod_dendro_ui(id)

mod_dendro_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
