% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_donneessup.R
\name{mod_donneessup_ui}
\alias{mod_donneessup_ui}
\alias{mod_donneessup_server}
\title{mod_donneessup_ui and mod_donneessup_server}
\usage{
mod_donneessup_ui(id)

mod_donneessup_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
