% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_incertitude.R
\name{mod_incertitude_ui}
\alias{mod_incertitude_ui}
\alias{mod_incertitude_server}
\title{mod_incertitude_ui and mod_incertitude_server}
\usage{
mod_incertitude_ui(id)

mod_incertitude_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
