% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_outputfiles.R
\name{mod_outputfiles_ui}
\alias{mod_outputfiles_ui}
\alias{mod_outputfiles_server}
\title{mod_outputfiles_ui and mod_outputfiles_server}
\usage{
mod_outputfiles_ui(id)

mod_outputfiles_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
