% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_outputoptions.R
\name{mod_outputoptions_ui}
\alias{mod_outputoptions_ui}
\alias{mod_outputoptions_server}
\title{mod_outputoptions_ui and mod_outputoptions_server}
\usage{
mod_outputoptions_ui(id)

mod_outputoptions_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
