% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_s2product.R
\name{mod_s2product_ui}
\alias{mod_s2product_ui}
\alias{mod_s2product_server}
\title{mod_s2product_ui and mod_s2product_server}
\usage{
mod_s2product_ui(id)

mod_s2product_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
